bug fixed wumanber

+++ src/algorithms.h
@@ -1,5 +1,5 @@
 
-int ShiftOr(char* pattern, char* text, long** C);
+long ShiftOr(char* pattern, char* text, long** C);
 long** buildMasks(char* pattern);
-int WuManber(char* pattern, char* text, long** C, int r);
+long WuManber(char* pattern, char* text, long** C, int r);
 // void buildMasks(long** C, int c_size, char* pattern, int p_size, int remain_bits);
